batch operate

+++ PublicCMS-admin/src/com/publiccms/admin/views/controller/cms/CmsTagController.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.publiccms.entities.cms.CmsTag;
+import com.publiccms.logic.service.cms.CmsContentTagService;
 import com.publiccms.logic.service.cms.CmsTagService;
 import com.sanluan.common.base.BaseController;
 @Controller
@@ -14,6 +15,8 @@ import com.sanluan.common.base.BaseController;
 public class CmsTagController extends BaseController {
 	@Autowired
 	private CmsTagService service;
+	@Autowired
+	private CmsContentTagService contentTagService;
 
 	@RequestMapping(value = { "save" })
 	public String save(CmsTag entity, HttpServletRequest request) {
@@ -26,8 +29,13 @@ public class CmsTagController extends BaseController {
 	}
 
 	@RequestMapping(value = { "delete" })
-	public String delete(Integer id) {
-		service.delete(id);
+	public String delete(Integer[] ids) {
+		if (notEmpty(ids)) {
+			for (Integer id : ids) {
+				service.delete(id);
+				contentTagService.delete(id, null);
+			}
+		}
 		return "common/ajaxDone";
 	}
 }
